calculating the pre-discount charge, the discount amount, and final charge; added a test decorator for RentalAgreement class and made all member vars private to preserve encapsulation; added method to print rental agreement using a string builder.

+++ src/CheckOut.java
@@ -5,14 +5,14 @@ import java.sql.Connection;
 import java.sql.DriverManager;
 import java.sql.SQLException;
 import java.time.LocalDate;
-import java.util.Calendar;
 
 public class CheckOut {
     RentalAgreement rentalAgreement;
     final String toolCode;
     final int rentalDayCount;
-    final float discountPercentage;
+    final int discountPercentage;
     final LocalDate checkOutDate;
+
 
     public CheckOut(String toolCode, int rentalDayCount, int discountPercentage, LocalDate checkOutDate) throws Exception {
 
@@ -50,7 +50,7 @@ public class CheckOut {
     }
 
     // validate that the rental day count is at least one or greater
-    // if not throw back up to calling method so it can message user properly.
+    // if not throw back up to calling method, so it can message user properly.
     public void validateRentalDayCount(int rentalDayCount) throws Exception {
         if (rentalDayCount < 1) {
             throw new Exception("Rental day count is not 1 or greater. Please enter a value for number of rental days that is at least 1.");
